requirements.txt
aemeasure~=0.2.9
chardet>=4.0.0
conan>=2.0.0
gurobipy~=10.0.1
networkx>=2.5.1
pytest>=7.1.2
requests>=2.25.1
rich>=12.4.1
scikit-build>=0.16.7
setuptools>=61.2.0
slurminade~=0.5.2
